Modify initial page render
Added citations in README file to credit the creators of the images
used. Added images to the project in order to style the page. Removed
footer from source html. Added styling to css in order to make the page
look better. Added page.js module to hold logic for rendering the
initial page with the images attached.