Add main container content logic
Added logic to page.js that creates the basic layout for the main
section of the page. This includes functionality for creating a title
for the project, a sub-heading for the task section, and individual
tasks with checkboxes. Added styling for the checkbox container in order
to override the default ugly styling. Also added styling to other parts
of the main section to make everything space out neatly and align well.