Add task storage logic and styling
Added log.js file to the project and wrote functions to create tasks
and store them in an array. Altered styling for the checkboxes on the
page in order to keep the label text separate from the checkbox itself
so that it can be used for other functionality later in the project.