Add dynamic task render functionality
Added functionality to log.js and page.js so that the tasks on the page
would render dynamically based on the inputs from the task array.
Altered some styling of elements to make them align better on the page.
Changed 'new project' section in the header to be wrapped in a button
tag for later function targeting.