Add todo funcs
Added functions to the log.js file that create todo's and add them to
an array. Created another function that renders the todo's in the array
on the page based on what task section they go in. Altered page.js to
have a helper function that does the rendering of the todo's based on
the arguments from the call in log.js. Added styling so that the todo's
would show up one after another and line up with each other.