Add dynamic project funcs
Added funcs to the log.js file that create new projects and add them to
the project array. Created funcs in log.js and page.js that render the
projects to the page based on what section they are assigned to.
Adjusted styling of all the projects so that they would line up and look
neater on the page.