Add list generation funcs and button functionality
Added funcs to log.js that dynamically create lists and store them in
the list array. Also added func in page.js that displays the lists on
the page. Adjusted plus elements to be wrapped in button tags and
attached event listeners to each so that they would call their
respective creation and rendering funcs in order to create lists and
projects more dynamically. Added styling to the sidebar so that the
newly rendered content would load with the correct formatting.