Add new todo dropdown functionality and styling
Added function in page.js that dynamically renders and ellipsis on each
sub-section that shows a dropdown when clicked. Added button to that
dropdown that adds a new todo to the sub-section it is attached to.
Added styling to the dropdown and the new todo button to make them look
better when rendered on the page.