Add project and list deletion funcs
Added functions to page.js that delete the list or project specified
based on the context of which button was pressed and what project or
list it is attached to. Added warning in log.js that checks if there
aren't any lists and the user tries to make a project. Added styling
to the new dropdowns for the lists and the projects.